src/requirements.txt
pytz
boto3
aws-xray-sdk
amazon-dax-client